backend/requirements.txt
absl-py==2.1.0
aiomysql==0.2.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-asgi-testclient==1.4.11
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
blinker==1.4
boto3==1.34.103
botocore==1.34.103
Brlapi==0.8.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
chromedriver-py==124.0.6367.201
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.1
cookiecutter==1.7.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==19.16.0
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
Faker==24.14.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-sessions==0.3.2
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.13.4
fire==0.6.0
fonttools==4.51.0
fsspec==2024.3.1
future==0.18.2
gino==1.0.1
gino-starlette==0.1.6
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.63.0
gunicorn==22.0.0
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib-metadata==4.6.4
infinity==1.5
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jinja2-time==0.2.0
jmespath==1.0.1
jsonschema==3.2.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.5
manage-fastapi==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.5.0
poyo==0.5.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-bcrypt==0.4
pyasn1==0.6.0
PyAutoGUI==0.9.54
pybboxes==0.1.6
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.15
pydantic-settings==2.2.1
pydantic_core==2.18.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.2.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.0.3
questionary==1.10.0
reportlab==3.6.8
requests==2.32.2
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
sahi==0.11.15
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.20.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.41
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
starlette-session==0.4.3
sympy==1.12
systemd-python==234
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
text-unidecode==1.3
texttable==1.6.4
thop==0.1.1.post2209072238
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traces==0.6.3
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
triton==2.3.0
ttystatus==0.38
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.2.3
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarg==0.1.9
yolov5==7.0.9
zipp==1.0.0
